JobMaster/requirements.txt
openai==1.12.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-openai==0.0.5
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6 